toolbox_v3/requirements.txt
groq
python-dotenv
pyautogui
Pillow
pytesseract
pyobjc-framework-ApplicationServices
pyobjc-framework-Cocoa
pyobjc-framework-Quartz
opencv-python
numpy
supabase
pyperclip
